more xcache info and new example
Changed the xcache note to be a whole section, though it could still use a little more information from somebody more familiar with its details. I have also changed the opendata example this section used to have to one which uses CMS data and is a little more physically interesting.